+++ frontend/src/cards/TableCard.tsx
@@ -37,7 +37,7 @@ import type { CountColsMap } from '../charts/mapGlobals'
 import HetNotice from '../styles/HetComponents/HetNotice'
 import { generateSubtitle } from '../charts/utils'
 import HetDivider from '../styles/HetComponents/HetDivider'
-import { sortByIncome } from '../data/sorting/sortingUtils'
+import { sortForVegaByIncome } from '../data/sorting/IncomeSorterStrategy'
 
 // We need to get this property, but we want to show it as
 // part of the "population_pct" column, and not as its own column
@@ -166,7 +166,7 @@ export default function TableCard(props: TableCardProps) {
           data.length <= 0
 
         if (props.demographicType === 'income') {
-          data = sortByIncome(data)
+          data = sortForVegaByIncome(data)
         }
 
         return (
