+++ frontend/src/charts/disparityBarChart/Chart.tsx
@@ -21,7 +21,7 @@ import {
 } from '../utils'
 import type { MetricConfig } from '../../data/config/MetricConfig'
 import { DEMOGRAPHIC_DISPLAY_TYPES_LOWER_CASE } from '../../data/query/Breakdowns'
-import { sortByIncome } from '../../data/sorting/sortingUtils'
+import { sortForVegaByIncome } from '../../data/sorting/IncomeSorterStrategy'
 
 export const altLightMetric: MetricConfig = {
   chartTitle: 'Population Share (ACS)',
@@ -69,7 +69,7 @@ export function DisparityBarChart(props: DisparityBarChartProps) {
   }
 
   if (props.demographicType === 'income') {
-    dataFromProps = sortByIncome(dataFromProps)
+    dataFromProps = sortForVegaByIncome(dataFromProps)
   }
 
   // add delimiter for line breaks in column axis labels
