+++ frontend/src/data/sorting/IncomeSorterStrategy.ts
@@ -2,6 +2,28 @@ import type { Breakdowns } from '../query/Breakdowns'
 import type { Row } from '../utils/DatasetTypes'
 import { AbstractSortStrategy } from './AbstractDataSorter'
 
+export const regexStripIncomeString = /^[^\d-]*(\d+)/
+
+export function compareIncome(a: Row, b: Row): number {
+  const aValue = a?.['income']?.match(regexStripIncomeString)?.[1] || 0
+  const bValue = b?.['income']?.match(regexStripIncomeString)?.[1] || 0
+  return aValue - bValue
+}
+
+export function sortForVegaByIncome(data: Row[]): Row[] {
+  const dataWithUnder = data.filter((row: Row) => {
+    return row['income'].includes('Under') || row['income'].includes('All')
+  })
+  const dataWithoutUnder = data.filter((row: Row) => {
+    return !row['income'].includes('Under') && !row['income'].includes('All')
+  })
+
+  dataWithUnder.sort(compareIncome)
+  dataWithoutUnder.sort(compareIncome)
+
+  return [...dataWithUnder, ...dataWithoutUnder]
+}
+
 export class IncomeSorterStrategy extends AbstractSortStrategy {
   frontValues: string[]
   backValues: string[]
