+++ frontend/src/data/sorting/sortingUtils.ts
@@ -19,25 +19,3 @@ export function getSortArgs(
       return []
   }
 }
-
-export const regexStripIncomeString = /^[^\d-]*(\d+)/
-
-export function compareIncome(a: Row, b: Row): number {
-  const aValue = a?.['income']?.match(regexStripIncomeString)?.[1] || 0
-  const bValue = b?.['income']?.match(regexStripIncomeString)?.[1] || 0
-  return aValue - bValue
-}
-
-export function sortByIncome(data: Row[]): Row[] {
-  const dataWithUnder = data.filter((row: Row) => {
-    return row['income'].includes('Under') || row['income'].includes('All')
-  })
-  const dataWithoutUnder = data.filter((row: Row) => {
-    return !row['income'].includes('Under') && !row['income'].includes('All')
-  })
-
-  dataWithUnder.sort(compareIncome)
-  dataWithoutUnder.sort(compareIncome)
-
-  return [...dataWithUnder, ...dataWithoutUnder]
-}
